refactor progress and steps into components
